Code refactored. Final Commit.
